pyproject.toml
# Install dependencies listed in pyproject.toml
# $ uv sync --extra dev
# Update all dependencies to the latest versions
# $ uv lock --upgrade && uv sync --extra dev
# Add a new package as a dependency to your project
# $ uv add <package>
# Add a new package as a development dependency
# $ uv add <package> --dev
# Remove a package
# $ uv remove <package>
[project]
name = "article_rss"
version = "0.1.0"
description = ""
requires-python = ">=3.14"

authors = [
    { name = "acd1034", email = "48613285+acd1034@users.noreply.github.com" },
]

dependencies = [
    "feedgen>=1.0.0",
    "feedparser>=6.0.12",
    "google-genai>=1.68.0",
    "joblib>=1.5.3",
    "matplotlib>=3.5.2",
    "numpy>=2.0.0",
    "pytest>=8.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.1",
    "requests>=2.32.5",
    "rich>=13.7.1",
    "tellurium",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests",
]

[project.optional-dependencies]
dev = ["ruff>=0.9.8", "mypy>=1.13.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
tellurium = { git = "https://github.com/acd1034/tellurium.git", rev = "c587dcace28a654a279d30c9cf38811490cad0b7" }

[tool.ruff]
# 1行の最大文字数 (default: 88)
line-length = 88

[tool.ruff.lint]
# チェックするエラーの種類
# https://docs.astral.sh/ruff/rules/
select = [
    "F",  # pyflakes
    "I",  # isort
    "E",  # pycodestyle error
    "W",  # pycodestyle warning
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
# 除外するエラーの種類
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long
    "F503",  # line break before binary operator
    "UP007", # Use `X | Y` for type annotations
    "UP045", # Use `X | None` for type annotations
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
# [tool.ruff.lint.isort]
# known-third-party = ["fastapi", "pydantic", "starlette"]

# mypy の設定
[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
